In progress: re-writing whole program.
I wanted to add the capability to save the image WCS information to
the output postage stamps of TIFAA. But then I noticed I can also add
a multi-thread option and I noticed that the whole management of TIFAA
(being one of the first programs I wrote) is very primitive. So I
began the job of re-writing it based on my new methods while also
adding the two mentioned capabilities.

Currently the user-interface (in ui.c and ui.h) is complete.