Re-written old algorithm.
The old algorithm is now completely re-written based on the pthreads,
the only problem is that wcslib uses a non-thread-safe method to parse
through the headers and so, I had to put a mutex on that part of the
code that slows down the job when more threads are used!!! But it
still works fast when only one thread is used.

I hope wcslib uses a thread-safe method soon.