WCS coordinates added, information included in FITS header.
The main update now is that every output FITS image has WCS
information in it, activating RA and Dec information for all its
pixels.

I also added some comments to the FITS header that will tell the
reader what RA and Dec were used to create this postage stamp, along
with the postage stamp size and resolution.

The README file was also updated.