Option to multiply with survey weight images.
I added an option to multiply each region that is pulled out of the
survey image into the weight image (provided by the user). This will
make the job of those who want to work on counts based pixels easier,
especially when the input science images are in counts/second.