Correct help and copyright information.
In the copyright information on top of each source file there was a
typo: tiffa instead of tifaa! So I corrected it. Also in printing
the help, I noticed that I had mistakenly used a wrong name for
tifaa. So I corrected that too.

I had also forgot to explain in the help that when using wildcards a
`\` is necessary before the `*` so the shell doesn't expand it. I
added the explanation.